build(run): initial package
Run files are more portable than AppImages thanks to their lack of dependency on FUSE. Plus they aren't made by the AppImage Community who continue to make questionable decisions. These .run files serve as a alternative for anyone who wants them. I will integrate self updating soon, same as the tar.gz package. While some Linux/FreeBSD users may not know what a `.run` file is, it is what the NVIDIA drivers use along with Davinci Resolve on Linux.

To run it, simply mark it as executable and you're off to the races.

In a terminal, it can be used like so:

`Rokon-1.0.0-x86_64.run -- "--version"`

The `--` is important so the .run file knows you want to pass arguments to Rokon, not the self extracting script.