build(run): use self extract instead of makeself
! IMPORTANT!  !

Commit 2c5ce91606c5c91e624881189bb53eac13f4a025 's mentioned usage has changed.

makeself changed the command line usage by forcing you to use `--`  and was marked as a shell script instead of a executable and I ran it through shell script compilers and it just couldn't be helped. Additionally, a zsync file is generated now for the run file to use later once I figure out how to implement part of the zsync protocol into Rokon.

Normally, switching the usage of the command line would be a breaking change but since we're in a pre-release 0.0.0 phase (despite the VERSION) file so it's just a simple build change.

Down with AppImages!

Finally: Keep it Golang, loser!